Adds case for single instance EB deployments
The previous configuration would break on single instance EB deployments since, without a load balancer, x-forwarded-proto would not be set leading to an infinite redirect loop.  This adds a check for whether or not x-forwarded-proto is set, and only redirects if it is.  One conceivable use case for this is an EB application that has a single server (no load balancer or https) deployment for testing or demos and a multi-server deployment for production.